Add CSS to eggfriedrice.html
The website was looking bland with no styles at all. We wouldn't want it to go out of style.

One problem I didn't really look into was that I couldn't make all  of the elments align center properly so I have to use numbers to make it look center. It doesn't function the best way so in the future, try to fix it.